requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0,<2.0.0
pyjwt>=2.10.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
werkzeug>=3.0.0,<4.0.0

discord.py>=2.3.0,<3.0.0
aiohttp>=3.9.0
pillow>=10.0.0
